feat(directory): implement DirectoryScreen with filtering and modal functionality
- Add DirectoryScreen component for displaying places with search and category filtering.
- Create BudgetSlider component for setting price limits.
- Implement CategoryFilter component for selecting place categories.
- Add DirectoryHeader for search input and location display.
- Create PlaceCard component to display individual place details.
- Implement PlaceDetailModal for showing detailed information about selected places.
- Add styles for all components to ensure consistent UI.
