+++ client/features/directory/components/DirectoryHeader.tsx
@@ -1,0 +1,49 @@
+import React from "react";
+import { View, Text, TextInput, TouchableOpacity } from "react-native";
+import { Ionicons } from "@expo/vector-icons";
+import { styles } from "../styles/directory.styles";
+
+interface DirectoryHeaderProps {
+  searchQuery: string;
+  setSearchQuery: (query: string) => void;
+  handleSearch: () => void;
+}
+
+export const DirectoryHeader = ({
+  searchQuery,
+  setSearchQuery,
+  handleSearch,
+}: DirectoryHeaderProps) => {
+  return (
+    <View style={styles.header}>
+      <View style={styles.headerTop}>
+        <View style={styles.locationContainer}>
+          <Ionicons name="location" size={20} color="#fff" />
+          <Text style={styles.locationText}>Cavite State University</Text>
+        </View>
+        <TouchableOpacity>
+          <Ionicons name="heart-outline" size={28} color="#fff" />
+        </TouchableOpacity>
+      </View>
+
+      {/* Search Bar */}
+      <View style={styles.searchContainer}>
+        <Ionicons name="search" size={20} color="#999" />
+        <TextInput
+          style={styles.searchInput}
+          placeholder="Search for cafés, restaurants, and launderettes"
+          placeholderTextColor="#999"
+          value={searchQuery}
+          onChangeText={setSearchQuery}
+          onSubmitEditing={handleSearch}
+          returnKeyType="search"
+        />
+        {searchQuery.length > 0 && (
+          <TouchableOpacity onPress={() => setSearchQuery("")}>
+            <Ionicons name="close-circle" size={20} color="#999" />
+          </TouchableOpacity>
+        )}
+      </View>
+    </View>
+  );
+};
