+++ client/features/directory/styles/directory.styles.ts
@@ -1,0 +1,338 @@
+import { StyleSheet } from "react-native";
+
+export const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: "#f5f5f5",
+  },
+  header: {
+    backgroundColor: "#FF6B35",
+    paddingTop: 50,
+    paddingHorizontal: 20,
+    paddingBottom: 30,
+  },
+  headerTop: {
+    flexDirection: "row",
+    justifyContent: "space-between",
+    alignItems: "center",
+    marginBottom: 15,
+  },
+  locationContainer: {
+    flexDirection: "row",
+    alignItems: "center",
+    gap: 5,
+  },
+  locationText: {
+    color: "#fff",
+    fontSize: 14,
+    fontWeight: "600",
+  },
+  searchContainer: {
+    flexDirection: "row",
+    alignItems: "center",
+    backgroundColor: "#fff",
+    borderRadius: 10,
+    paddingHorizontal: 15,
+    paddingVertical: 12,
+    gap: 10,
+  },
+  searchInput: {
+    flex: 1,
+    fontSize: 14,
+    color: "#333",
+  },
+  limitSection: {
+    flex: 1,
+    paddingHorizontal: 20,
+    paddingVertical: 20,
+    backgroundColor: "#ffffffff",
+    marginTop: -17,
+    borderTopLeftRadius: 20,
+    borderTopRightRadius: 20,
+    shadowColor: "#000",
+    shadowOffset: { width: 0, height: 2 },
+    shadowOpacity: 0.1,
+    shadowRadius: 5,
+    elevation: 3,
+  },
+  limitLabel: {
+    fontSize: 14,
+    fontWeight: "600",
+    color: "#000000ff",
+    marginBottom: 10,
+  },
+  limitContainer: {
+    flexDirection: "row",
+    alignItems: "center",
+    gap: 15,
+  },
+  sliderContainer: {
+    flex: 1,
+  },
+  slider: {
+    width: "100%",
+    height: 40,
+  },
+  limitBadge: {
+    backgroundColor: "#fff",
+    borderWidth: 2,
+    borderColor: "#FF6B35",
+    borderRadius: 8,
+    paddingHorizontal: 12,
+    paddingVertical: 6,
+  },
+  limitText: {
+    color: "#FF6B35",
+    fontWeight: "700",
+    fontSize: 14,
+  },
+  categoryContainer: {
+    paddingHorizontal: 20,
+    paddingVertical: 20,
+    gap: 10,
+  },
+  categoryTab: {
+    paddingHorizontal: 20,
+    paddingVertical: 10,
+    borderRadius: 8,
+    borderWidth: 1,
+    borderColor: "#ddd",
+    backgroundColor: "#fff",
+    marginRight: 10,
+  },
+  categoryTabActive: {
+    backgroundColor: "#FF6B35",
+    borderColor: "#FF6B35",
+  },
+  categoryText: {
+    fontSize: 14,
+    fontWeight: "500",
+    color: "#666",
+  },
+  categoryTextActive: {
+    color: "#fff",
+  },
+  resultsCount: {
+    paddingHorizontal: 20,
+    paddingBottom: 10,
+  },
+  resultsText: {
+    fontSize: 14,
+    color: "#666",
+    fontWeight: "600",
+  },
+  cardContainer: {
+    paddingHorizontal: 20,
+    marginBottom: 15,
+  },
+  card: {
+    backgroundColor: "#fff",
+    borderRadius: 15,
+    overflow: "hidden",
+    shadowColor: "#000",
+    shadowOffset: { width: 0, height: 2 },
+    shadowOpacity: 0.1,
+    shadowRadius: 5,
+    elevation: 3,
+  },
+  cardImage: {
+    width: "100%",
+    height: 200,
+    backgroundColor: "#e0e0e0",
+  },
+  cardBadge: {
+    position: "absolute",
+    top: 10,
+    right: 10,
+    backgroundColor: "#FF6B35",
+    paddingHorizontal: 12,
+    paddingVertical: 6,
+    borderRadius: 8,
+  },
+  cardBadgeText: {
+    color: "#fff",
+    fontWeight: "700",
+    fontSize: 12,
+  },
+  cardContent: {
+    padding: 15,
+  },
+  categoryChip: {
+    backgroundColor: "#FF6B35",
+    alignSelf: "flex-start",
+    paddingHorizontal: 12,
+    paddingVertical: 6,
+    borderRadius: 6,
+    marginBottom: 8,
+  },
+  categoryChipText: {
+    color: "#fff",
+    fontSize: 11,
+    fontWeight: "600",
+  },
+  cardTitle: {
+    fontSize: 18,
+    fontWeight: "700",
+    color: "#333",
+    marginBottom: 4,
+  },
+  cardLocation: {
+    fontSize: 13,
+    color: "#666",
+    marginBottom: 8,
+  },
+  tagsContainer: {
+    flexDirection: "row",
+    flexWrap: "wrap",
+    gap: 6,
+  },
+  tag: {
+    backgroundColor: "#f0f0f0",
+    paddingHorizontal: 10,
+    paddingVertical: 4,
+    borderRadius: 12,
+  },
+  tagText: {
+    fontSize: 11,
+    color: "#666",
+  },
+  emptyState: {
+    alignItems: "center",
+    paddingVertical: 60,
+    paddingHorizontal: 40,
+  },
+  emptyTitle: {
+    fontSize: 18,
+    fontWeight: "600",
+    color: "#333",
+    textAlign: "center",
+    marginTop: 20,
+    marginBottom: 8,
+  },
+  emptySubtitle: {
+    fontSize: 14,
+    color: "#999",
+    textAlign: "center",
+  },
+  modalContainer: {
+    flex: 1,
+    backgroundColor: "#f5f5f5",
+  },
+  modalImageContainer: {
+    position: "relative",
+    height: 250,
+  },
+  modalImage: {
+    width: "100%",
+    height: "100%",
+  },
+  modalBackButton: {
+    position: "absolute",
+    top: 50,
+    left: 20,
+    width: 40,
+    height: 40,
+    borderRadius: 20,
+    backgroundColor: "rgba(0,0,0,0.3)",
+    justifyContent: "center",
+    alignItems: "center",
+  },
+  modalHeartButton: {
+    position: "absolute",
+    top: 50,
+    right: 20,
+    width: 40,
+    height: 40,
+    borderRadius: 20,
+    backgroundColor: "rgba(0,0,0,0.3)",
+    justifyContent: "center",
+    alignItems: "center",
+  },
+  modalContent: {
+    flex: 1,
+  },
+  modalHeader: {
+    backgroundColor: "#fff",
+    padding: 20,
+    flexDirection: "row",
+    justifyContent: "space-between",
+    alignItems: "flex-start",
+  },
+  modalTitle: {
+    fontSize: 24,
+    fontWeight: "700",
+    color: "#333",
+    marginBottom: 4,
+  },
+  modalLocation: {
+    fontSize: 14,
+    color: "#666",
+  },
+  modalPriceBadge: {
+    backgroundColor: "#FF6B35",
+    paddingHorizontal: 12,
+    paddingVertical: 8,
+    borderRadius: 8,
+  },
+  modalPriceText: {
+    color: "#fff",
+    fontWeight: "700",
+    fontSize: 13,
+  },
+  menuSection: {
+    backgroundColor: "#fff",
+    marginTop: 10,
+    padding: 20,
+  },
+  menuTitle: {
+    fontSize: 20,
+    fontWeight: "700",
+    color: "#333",
+    marginBottom: 20,
+  },
+  menuItem: {
+    flexDirection: "row",
+    marginBottom: 20,
+    gap: 15,
+  },
+  menuItemIcon: {
+    width: 60,
+    height: 60,
+    backgroundColor: "#FF6B35",
+    borderRadius: 8,
+  },
+  menuItemContent: {
+    flex: 1,
+  },
+  menuItemName: {
+    fontSize: 16,
+    fontWeight: "600",
+    color: "#333",
+    marginBottom: 4,
+  },
+  menuItemMeta: {
+    fontSize: 12,
+    color: "#999",
+    marginBottom: 4,
+  },
+  menuItemPrice: {
+    fontSize: 18,
+    fontWeight: "700",
+    color: "#333",
+  },
+  noMealsContainer: {
+    alignItems: "center",
+    paddingVertical: 40,
+  },
+  noMealsText: {
+    fontSize: 16,
+    fontWeight: "600",
+    color: "#666",
+    marginBottom: 8,
+  },
+  noMealsSubtext: {
+    fontSize: 14,
+    color: "#999",
+    textAlign: "center",
+  },
+});
